distincion de roles

+++ app/Http/Controllers/UsersCtrl.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\User;
 use Illuminate\Support\Facades\Hash;
+use Illuminate\Support\Facades\Auth;
 
 class UsersCtrl extends Controller
 {
@@ -15,8 +16,11 @@ class UsersCtrl extends Controller
      */
     public function index()
     {
-        $usuarios = User::orderByDesc('fechaalta')->paginate(4);
-        return view('usuarios.usuariosVer', compact('usuarios'));
+        if (Auth::user()->tipo == 'administrador') {
+            $usuarios = User::orderByDesc('fechaalta')->paginate(4);
+            return view('usuarios.usuariosVer', compact('usuarios'));
+        } else
+            return redirect()->action([AuthenticatedSessionController::class, 'destroy']);
     }
 
     /**
@@ -26,7 +30,10 @@ class UsersCtrl extends Controller
      */
     public function create()
     {
-        return view('usuarios.usuarioCrear');
+        if (Auth::user()->tipo == 'administrador') {
+            return view('usuarios.usuarioCrear');
+        } else
+            return redirect()->action([AuthenticatedSessionController::class, 'destroy']);
     }
 
     /**
@@ -37,33 +44,36 @@ class UsersCtrl extends Controller
      */
     public function store(Request $request)
     {
-        // Almaceno la contraseña para comprobar que coincide con su repeticion en la validacion
-        $pass = $request->password;
-        $datos = $request->validate([
-            'name' => ['required', 'max:255', 'alpha_num'],
-            'email' => ['required', 'max:255', 'regex:/^[\w\-\.]+@([\w\-]+\.)+[\w\-]{2,4}$/'],
-            'password' => ['required', 'max:255'],
-            'passrep' => [
-                'required', 'max:255',
-                function ($attribute, $value, $fail) use ($pass) {
-                    if ($value != $pass) {
-                        $fail('Las contraseñas no coinciden.');
+        if (Auth::user()->tipo == 'administrador') {
+            // Almaceno la contraseña para comprobar que coincide con su repeticion en la validacion
+            $pass = $request->password;
+            $datos = $request->validate([
+                'name' => ['required', 'max:255', 'alpha_num'],
+                'email' => ['required', 'max:255', 'regex:/^[\w\-\.]+@([\w\-]+\.)+[\w\-]{2,4}$/'],
+                'password' => ['required', 'max:255'],
+                'passrep' => [
+                    'required', 'max:255',
+                    function ($attribute, $value, $fail) use ($pass) {
+                        if ($value != $pass) {
+                            $fail('Las contraseñas no coinciden.');
+                        }
                     }
-                }
-            ],
-            'dni' => ['required', 'max:45', 'regex:/((^[A-Z]{1}[0-9]{7}[A-Z0-9]{1}$|^[T]{1}[A-Z0-9]{8}$)|^[0-9]{8}[A-Z]{1}$)/'],
-            'telefono' => ['required', 'max:45', 'regex:/(\+34|0034|34)?[ -]*(6|7|8|9)[ -]*([0-9][ -]*){8}/'],
-            'direccion' => ['required', 'max:45'],
-            'tipo' => ['required', 'max:45'],
-        ]);
-        // Encripto la contraseña para poder utilizar el login de Breeze
-        $datos['password'] = Hash::make($request->password);
-        // Almaceno la fecha del momento de creación del usuario para mostrarla como fecha de alta
-        $datos['fechaalta'] = date("Y-m-d\TH:i");
-        // Elimino el campo passrep del array de datos ya que no es necesario insertarlo en la base de datos
-        unset($datos['passrep']);
-        $usuario = User::create($datos);
-        return view('usuarios.usuarioVerDetalles', compact('usuario'));
+                ],
+                'dni' => ['required', 'max:45', 'regex:/((^[A-Z]{1}[0-9]{7}[A-Z0-9]{1}$|^[T]{1}[A-Z0-9]{8}$)|^[0-9]{8}[A-Z]{1}$)/'],
+                'telefono' => ['required', 'max:45', 'regex:/(\+34|0034|34)?[ -]*(6|7|8|9)[ -]*([0-9][ -]*){8}/'],
+                'direccion' => ['required', 'max:45'],
+                'tipo' => ['required', 'max:45'],
+            ]);
+            // Encripto la contraseña para poder utilizar el login de Breeze
+            $datos['password'] = Hash::make($request->password);
+            // Almaceno la fecha del momento de creación del usuario para mostrarla como fecha de alta
+            $datos['fechaalta'] = date("Y-m-d\TH:i");
+            // Elimino el campo passrep del array de datos ya que no es necesario insertarlo en la base de datos
+            unset($datos['passrep']);
+            $usuario = User::create($datos);
+            return view('usuarios.usuarioVerDetalles', compact('usuario'));
+        } else
+            return redirect()->action([AuthenticatedSessionController::class, 'destroy']);
     }
 
     /**
@@ -74,8 +84,11 @@ class UsersCtrl extends Controller
      */
     public function show($id)
     {
-        $usuario = User::find($id);
-        return view('usuarios.usuarioVerDetalles', compact('usuario'));
+        if (Auth::user()->tipo == 'administrador' || $id == Auth::user()->id) {
+            $usuario = User::find($id);
+            return view('usuarios.usuarioVerDetalles', compact('usuario'));
+        } else
+            return redirect()->action([AuthenticatedSessionController::class, 'destroy']);
     }
 
     /**
@@ -86,8 +99,11 @@ class UsersCtrl extends Controller
      */
     public function edit($id)
     {
-        $usuario = User::find($id);
-        return view('usuarios.usuarioModificar', compact('usuario'));
+        if (Auth::user()->tipo == 'administrador' || $id == Auth::user()->id) {
+            $usuario = User::find($id);
+            return view('usuarios.usuarioModificar', compact('usuario'));
+        } else
+            return redirect()->action([AuthenticatedSessionController::class, 'destroy']);
     }
 
     /**
@@ -99,28 +115,31 @@ class UsersCtrl extends Controller
      */
     public function update(Request $request, $id)
     {
-        // Almaceno la contraseña para comprobar que coincide con su repeticion en la validacion
-        $pass = $request->password;
-        $datos = $request->validate([
-            'name' => ['required', 'max:255', 'alpha_num'],
-            'email' => ['required', 'max:255', 'regex:/^[\w\-\.]+@([\w\-]+\.)+[\w\-]{2,4}$/'],
-            'password' => ['required', 'max:255'],
-            'passrep' => [
-                'required', 'max:255',
-                function ($attribute, $value, $fail) use ($pass) {
-                    if ($value != $pass) {
-                        $fail('Las contraseñas no coinciden.');
+        if (Auth::user()->tipo == 'administrador' || $id == Auth::user()->id) {
+            // Almaceno la contraseña para comprobar que coincide con su repeticion en la validacion
+            $pass = $request->password;
+            $datos = $request->validate([
+                'name' => ['required', 'max:255', 'alpha_num'],
+                'email' => ['required', 'max:255', 'regex:/^[\w\-\.]+@([\w\-]+\.)+[\w\-]{2,4}$/'],
+                'password' => ['required', 'max:255'],
+                'passrep' => [
+                    'required', 'max:255',
+                    function ($attribute, $value, $fail) use ($pass) {
+                        if ($value != $pass) {
+                            $fail('Las contraseñas no coinciden.');
+                        }
                     }
-                }
-            ],
-            'dni' => ['required', 'max:45', 'regex:/((^[A-Z]{1}[0-9]{7}[A-Z0-9]{1}$|^[T]{1}[A-Z0-9]{8}$)|^[0-9]{8}[A-Z]{1}$)/'],
-            'telefono' => ['required', 'max:45', 'regex:/(\+34|0034|34)?[ -]*(6|7|8|9)[ -]*([0-9][ -]*){8}/'],
-            'direccion' => ['required', 'max:45'],
-            'fechaalta' => ['required', 'date_format:Y-m-d\TH:i'],
-            'tipo' => ['required', 'max:45'],
-        ]);
-        User::find($id)->update($datos);
-        return $this->show($id);
+                ],
+                'dni' => ['required', 'max:45', 'regex:/((^[A-Z]{1}[0-9]{7}[A-Z0-9]{1}$|^[T]{1}[A-Z0-9]{8}$)|^[0-9]{8}[A-Z]{1}$)/'],
+                'telefono' => ['required', 'max:45', 'regex:/(\+34|0034|34)?[ -]*(6|7|8|9)[ -]*([0-9][ -]*){8}/'],
+                'direccion' => ['required', 'max:45'],
+                'fechaalta' => ['required', 'date_format:Y-m-d\TH:i'],
+                'tipo' => ['required', 'max:45'],
+            ]);
+            User::find($id)->update($datos);
+            return $this->show($id);
+        } else
+            return redirect()->action([AuthenticatedSessionController::class, 'destroy']);
     }
 
     /**
@@ -131,14 +150,20 @@ class UsersCtrl extends Controller
      */
     public function destroy($id)
     {
-        $usuario = User::find($id);
-        $usuario->delete();
-        return view('usuarios.usuarioEliminado', ['id' => $id]);
+        if (Auth::user()->tipo == 'administrador') {
+            $usuario = User::find($id);
+            $usuario->delete();
+            return view('usuarios.usuarioEliminado', ['id' => $id]);
+        } else
+            return redirect()->action([AuthenticatedSessionController::class, 'destroy']);
     }
 
     public function confirmarBorrado($id)
     {
-        $usuario = User::find($id);
-        return view('usuarios.usuarioEliminar', compact('usuario'));
+        if (Auth::user()->tipo == 'administrador') {
+            $usuario = User::find($id);
+            return view('usuarios.usuarioEliminar', compact('usuario'));
+        } else
+            return redirect()->action([AuthenticatedSessionController::class, 'destroy']);
     }
 }
